requirements.txt
discord_py_interactions==5.11.0
interactions.py==5.14.0
llama_index==0.12.33
llama-index-embeddings-huggingface==0.1.4
llama-index-embeddings-openai==0.1.3
llama-index-llms-groq==0.1.2
numpy==2.2.5
python-dotenv==1.1.0
PyYAML==6.0.2
torch==2.2.2
sentence-transformers==2.6.1
pynacl==1.5.0
setuptools==69.5.1
tree-sitter=0.24.0
tree-sitter-python=0.23.6
tree-sitter-c-sharp=0.23.6
sentence_transformers==4.1.0
loguru==0.7.3